Move shared code to helper_funcs.sh for EWv1_CreateBuildRun scripts
The scripts initially had the same lines to add the vexec and usage functions,
parse command line arguments, and print info about the run. These now reside in
helper_funcs.sh. This is the only script that does not have execute permissions
set because it is never meant to run on its own.